# public/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TPLK004 - Sistem Absensi & Manajemen Akademik API",
    "description": "API Documentation untuk Sistem Absensi QR Code dengan Gamifikasi, Manajemen Tugas, dan Analytics. Sistem ini menyediakan endpoint untuk autentikasi, manajemen absensi, tugas, gamifikasi, dan berbagai fitur akademik lainnya.",
    "version": "1.0.0",
    "contact": {
      "name": "TPLK004 Support Team",
      "email": "admin@unpam.ac.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter token in format: Bearer {token}"
      },
      "sanctum": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Laravel Sanctum session authentication"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "object"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@unpam.ac.id"
          },
          "role": {
            "type": "string",
            "enum": ["mahasiswa", "dosen", "admin"],
            "example": "mahasiswa"
          },
          "avatar_url": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "points": {
            "type": "integer",
            "example": 1250
          },
          "level": {
            "type": "integer",
            "example": 5
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "description": "Authenticate user dan mendapatkan access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "role"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mahasiswa@unpam.ac.id"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["mahasiswa", "dosen", "admin"],
                    "example": "mahasiswa"
                  },
                  "remember": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Login berhasil"
                    },
                    "token": {
                      "type": "string",
                      "example": "1|abcdef123456..."
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Logout user",
        "description": "Revoke access token dan logout user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Logout berhasil"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/user": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get authenticated user",
        "description": "Mendapatkan data user yang sedang login",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User data retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}